+++ CHANGELOG.md
@@ -36,6 +36,7 @@ FEATURES:
   * Add support for stored shares, recovery parameters etc. during the initialization of Vault.
   * Supports the new fields (`hmac_accessor`, `jsonx` format etc.) for File and SysLog Audit Backends.
   * The `AWSGenerateDynamicCredentialsWithSecurityTokenAsync` method now supports the `timeToLive` parameter.
+  * The `Consul` backend now supports the listing functionality to roles `ConsulReadRoleListAsync`. (https://github.com/hashicorp/vault/issues/2065)
   * All the secret backends now support wrapping of the secret into a cubbyhole token. Wrapping support added for:
     * AWS Secret Backend
     * Cassandra Secret Backend